Handlers

+++ Subscriber.Services/Models/CardModel.cs
@@ -12,11 +12,11 @@ namespace Subscriber.Services.Models
 
         public string LastName { get; set; }
 
-        public int BMI { get; set; }
+        public float BMI { get; set; }
 
         public int Height { get; set; }
 
-        public int Weight { get; set; }
+        public decimal Weight { get; set; }
        
     }
